requirements.txt
# =============================================================================
# DEPENDENCIAS PRINCIPALES
# =============================================================================
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# =============================================================================
# VISUALIZACIÓN
# =============================================================================
plotly>=5.17.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4

# =============================================================================
# MANEJO DE ARCHIVOS
# =============================================================================
openpyxl>=3.1.0
xlrd>=2.0.1
chardet>=5.2.0

# =============================================================================
# SUPABASE CLIENT
# =============================================================================
supabase>=2.0.0
postgrest>=0.13.0

# =============================================================================
# UTILIDADES
# =============================================================================
python-dateutil>=2.8.2
pytz>=2023.3
